newsBookmarkServer/requirements.txt
argcomplete==1.9.3
asn1crypto==0.24.0
base58==0.2.4
boto3==1.9.152
botocore==1.12.152
certifi==2019.3.9
cffi==1.12.3
cfn-flip==1.2.0
chardet==3.0.4
Click==7.0
cryptography==2.6.1
Django==2.1.7
django-cors-headers==2.5.3
django-rest-auth==0.9.5
djangorestframework==3.9.3
djangorestframework-jwt==1.11.0
docutils==0.14
durationpy==0.5
future==0.16.0
futures==3.1.1
hjson==3.0.1
idna==2.8
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
mysqlclient==1.4.2.post1
placebo==0.9.0
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
pyOpenSSL==19.0.0
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2019.1
PyYAML==3.13
requests==2.22.0
s3transfer==0.2.0
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
tqdm==4.19.1
troposphere==2.4.7
Unidecode==1.0.23
urllib3==1.24.3
Werkzeug==0.15.4
wsgi-request-logger==0.4.6
zappa==0.47.1
